requirements.txt
absl-py==0.13.0
alabaster==0.7.12
alembic==1.6.5
appdirs==1.4.4
atari-py==0.2.6
attrs==21.2.0
Babel==2.9.1
black==21.7b0
box2d-py==2.3.8
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cliff==3.8.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
coverage==5.5
cycler==0.10.0
Cython==0.29.24
DataProperty==0.52.0
docutils==0.16
execnet==1.9.0
flake8==3.9.2
flake8-bugbear==21.4.3
google-auth==1.35.0
google-auth-oauthlib==0.4.5
greenlet==1.1.1
grpcio==1.39.0
gym==0.19.0
gym-minigrid==1.0.2
idna==3.2
imagesize==1.2.0
importlab==0.6.1
iniconfig==1.1.1
inputs==0.5
isort==5.9.3
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
libcst==0.3.20
livereload==2.6.3
llvmlite==0.37.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mbstrdecoder==1.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.2
ninja==1.10.2
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.3.56
optuna==2.9.1
packaging==21.0
panda-gym==1.1.0
pandas==1.3.2
pathspec==0.9.0
pathvalidate==2.4.1
pbr==5.6.0
Pillow==8.3.1
plotly==5.2.1
pluggy==0.13.1
prettytable==2.1.0
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyaml==21.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.7
pycodestyle==2.7.0
pyenchant==3.2.1
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
pytablewriter==0.62.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-env==0.6.2
pytest-forked==1.3.0
pytest-xdist==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
pytype==2021.8.11
pytz==2021.1
PyYAML==5.4.1
quadprog==0.1.8
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sb3-contrib==1.1.0
scikit-learn==0.24.2
scikit-optimize==0.8.1
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints==1.12.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-spelling==7.2.1
SQLAlchemy==1.4.23
stable-baselines3==1.1.0
stevedore==3.4.0
tabledata==1.2.0
tabulate==0.8.9
tcolorpy==0.1.1
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.1
torch==1.9.0
tornado==6.1
tqdm==4.62.1
typed-ast==1.4.3
typepy==1.2.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
